Changed Data storage to cloud

+++ backend/src/api/routes/ContentMap.js
@@ -1,37 +1,31 @@
 const { Router } = require('express');
 const { verifyUser } = require("../helpers/firebase");
 const fb = require("../helpers/firebase");
-
+const uuid = require("uuid");
+const oci = require("../helpers/oracle");
 
 const router = Router();
 
 
 /* Create a Ne content map */
-router.post("/", async (req, res) =>
-{
+router.post("/", async (req, res) => {
     const auth = req.headers.authorization;
     const token = auth.split(' ')[1];
 
     const { name, data } = req.body;
 
     // Check if the token exists
-    if (!token||!name) return res.status(400).json({ code: "AM101", error: "Missing token or name" });
+    if (!token || !name) return res.status(400).json({ code: "AM101", error: "Missing token or name" });
 
     //verify user
     const user = await verifyUser(token);
-    if (!user)
-    {
+    if (!user) {
         return res.status(403).json({ code: "AM102", error: "Invalid token" });
     }
 
-    // create a new content map as collection inside user's doc
-    const contentMap = {
-        name: name,
-        data: data ? JSON.stringify(data):"",
-        createdAt: fb.admin.firestore.FieldValue.serverTimestamp(),
-        updatedAt: fb.admin.firestore.FieldValue.serverTimestamp(),
-        Access: {[user.uid]: {role:"owner",email:user.email, name:user.name, type:"users"}}
-    }
+
+
+
 
     const db = fb.admin.firestore();
     const userRef = db.collection("users").doc(user.uid);
@@ -39,8 +33,25 @@ router.post("/", async (req, res) =>
     const doc = await userRef.get();
     if (!doc.exists) return res.status(404).json({ code: "AM107", error: "User not found" });
 
-
     const contentMapsRef = db.collection("contentMaps")
+
+    let DataId = uuid.v4();
+
+    // upload data to oracle cloud
+    const uploadData = await oci.AddData("B3", DataId, "application/json", JSON.stringify(data));
+
+    if (!uploadData.eTag) {
+        return res.status(500).json({ code: 500, error: "Uploading data failed" });
+    }
+
+    // create a new content map as collection inside user's doc
+    const contentMap = {
+        name: name,
+        data: DataId,
+        createdAt: fb.admin.firestore.FieldValue.serverTimestamp(),
+        updatedAt: fb.admin.firestore.FieldValue.serverTimestamp(),
+        Access: { [user.uid]: { role: "owner", email: user.email, name: user.name, type: "users" } }
+    }
 
     const contentMapRef = await contentMapsRef.add(contentMap);
     const contentMapId = contentMapRef.id;
@@ -62,11 +73,11 @@ router.get("/", async (req, res) => {
 
     // Check if the token exists
     if (!token) return res.status(400).json({ code: "AM101", error: "Missing token" });
-    
+
     //verify user
     const user = await verifyUser(token);
     if (!user) return res.status(403).json({ code: "AM102", error: "Invalid token" });
-    
+
 
     const db = fb.admin.firestore();
     const userRef = db.collection("users").doc(user.uid);
@@ -76,40 +87,14 @@ router.get("/", async (req, res) => {
 
     //get the content maps array of the user
     const contentMaps = doc.data().contentMaps;
-    
+
     return res.status(200).json(contentMaps);
 });
 
 
-/* Get a content map of a user */
-router.get("/:user/:id", async (req, res) => {
+/* Get a content map of a user - Please check dashboard API */
 
-    const token = req.headers?.authorization?.split(' ')[1];
-    const id = req.params.id;
-    const ResourceUser = req.params.user;
-
-    // Check if the token exists
-    if (!token || !ResourceUser) return res.status(400).json({ code: "AM101", error: "Missing token or user" });
-    
-    //verify user
-    const user = await verifyUser(token);
-    if (!user) return res.status(403).json({ code: "AM102", error: "Invalid token" });
-
-    const db = fb.admin.firestore();
-    const contentMap = await db.collection("contentMaps").doc(id).get();
-    
-
-    if (!contentMap.exists) return res.status(404).json({ code: "AM108", error: "Content map not found" });
-
-    const contentMapData = contentMap.data();
-
-    //check if user has access to the content map
-    if(!contentMapData.Access[user.uid]) return res.status(403).json({ code: "AM109", error: "User does not have access to the content map" });
-    
-    return res.status(200).json({...contentMapData,userAccess:contentMapData.Access[user.uid]?.role});
-});
-
-// untested
+/* Update a content map of a user */
 router.put("/:user/:id", async (req, res) => {
 
     const auth = req.headers.authorization;
@@ -117,11 +102,10 @@ router.put("/:user/:id", async (req, res) => {
     const token = auth?.split(' ')[1];
     const id = req.params.id;
 
-    const { name, data , share , access } = req.body;
+    const { name, data, share, access } = req.body;
 
-    // Check if the token exists
-    if (!token || !id || !resourceUser || (share && (share!=="view" || share!=="edit")))
-    {
+    // Check if the data exists
+    if (!token || !id || !resourceUser || (share && (share !== "view" || share !== "edit"))) {
         return res.status(400).json({ code: 400, error: "Missing token or id or resource user or share" });
     }
 
@@ -130,35 +114,48 @@ router.put("/:user/:id", async (req, res) => {
     if (!user) return res.status(403).json({ code: 403, error: "Invalid token" });
 
     const db = fb.admin.firestore();
-    
-    const contentMap =await db.collection("contentMaps").doc(id).get();
+
+    const contentMap = await db.collection("contentMaps").doc(id).get();
     if (!contentMap.exists) return res.status(404).json({ code: 404, error: "Content map not found" });
-    
+
 
     let contentMapData = contentMap.data();
-    if(!contentMapData.Access[user.uid]){
+    if (!contentMapData.Access[user.uid]) {
         return res.status(403).json({ code: 403, error: "User does not have access to the content map" });
     }
 
     let userRole = contentMapData.Access[user.uid].role;
-    
+
     let updatedContentMap = {
         ...contentMapData
     };
 
-    if(data && (userRole==="owner"|| userRole==="edit")){
-        updatedContentMap.data = JSON.stringify(data);
+    if (data && (userRole === "owner" || userRole === "edit")) {
+
+        // check if the data is uuid else generate new uuid
+        if (uuid.validate(contentMapData.data)) {
+            updatedContentMap.data = contentMapData.data;
+        } else {
+            updatedContentMap.data = uuid.v4();
+        }
+
+        // upload data to oracle cloud
+        const uploadData = await oci.AddData("B3", contentMapData.data, "application/json", JSON.stringify(data));
+
+        if (!uploadData.eTag) {
+            return res.status(500).json({ code: 500, error: "Uploading data failed" });
+        }
     }
 
-    if(share && userRole==="owner"){
+    if (share && userRole === "owner") {
         updatedContentMap.share = share;
     }
 
-    if(access && userRole==="owner"){
+    if (access && userRole === "owner") {
         updatedContentMap.Access = access;
     }
 
-    if(name && userRole==="owner"){
+    if (name && userRole === "owner") {
         updatedContentMap.name = name;
     }
 
@@ -166,10 +163,10 @@ router.put("/:user/:id", async (req, res) => {
     //set updated content map
     await db.collection("contentMaps").doc(id).set(updatedContentMap);
 
-    return res.status(200).json({ code:200, id: id });
+    return res.status(200).json({ code: 200, id: id });
 });
 
-// untested
+/* Delete a content map of a user */
 router.delete("/:id", async (req, res) => {
 
     const auth = req.headers.authorization;
@@ -178,18 +175,18 @@ router.delete("/:id", async (req, res) => {
 
     // Check if the token exists
     if (!token) return res.status(400).json({ code: 400, error: "Missing token" });
-    
+
 
     //verify user
     const user = await verifyUser(token);
     if (!user) return res.status(403).json({ code: 403, error: "Invalid token" });
-    
+
     const db = fb.admin.firestore();
-    const userRef =db.collection("users").doc(user.uid)
+    const userRef = db.collection("users").doc(user.uid)
     const doc = await userRef.get();
 
     if (!doc.exists) return res.status(404).json({ code: 404, error: "User not found" });
-    
+
     const contentMapsRef = db.collection("contentMaps");
     const contentMapRef = contentMapsRef.doc(id);
     const contentMap = await contentMapRef.get();
@@ -198,9 +195,14 @@ router.delete("/:id", async (req, res) => {
 
     // Ensure user is owner of the content map
     let contentMapData = contentMap.data();
-    if(!contentMapData.Access[user.uid] || contentMapData.Access[user.uid].role!=="owner"){
+    if (!contentMapData.Access[user.uid] || contentMapData.Access[user.uid].role !== "owner") {
         return res.status(403).json({ code: 403, error: "User does not have access to the operation" });
     }
+
+    // delete the data from oracle cloud
+    const deleteData = await oci.deleteData("B3", contentMapData.data);
+
+    if (!deleteData.lastModified) return res.status(500).json({ code: 500, error: "Deleting data failed" });
 
     await contentMapRef.delete();
 
@@ -211,7 +213,7 @@ router.delete("/:id", async (req, res) => {
         contentMaps: userContentMaps
     });
 
-    return res.status(200).json({ code:200, id: id });
+    return res.status(200).json({ code: 200, id: id });
 });
 
 
@@ -223,12 +225,12 @@ router.get("/search", async (req, res) => {
     const token = auth?.split(' ')[1];
     const query = req.query.query;
 
-    if(!query || query.length<3 || !token) return res.status(400).json({ code: 400, error: "Missing token or query or query length should be greater than 3" });
+    if (!query || query.length < 3 || !token) return res.status(400).json({ code: 400, error: "Missing token or query or query length should be greater than 3" });
 
     //verify user
     const user = await verifyUser(token);
     if (!user) return res.status(403).json({ code: "AM102", error: "Invalid token" });
-    
+
 
     const db = fb.admin.firestore();
     const usersRef = db.collection("users");
@@ -244,12 +246,12 @@ router.get("/search", async (req, res) => {
 
     //get teams based on name
     const teamsBasedOnName = await teamsRef.where("name", ">=", query).where("name", "<=", query + "\uf8ff").get();
-    
+
 
 
     // add them to users array, remove duplicates
     users = [...usersBasedOnEmail.docs, ...usersBasedOnUserId.docs];
-    const usersData = users.map(doc => ({ id: doc.id, email: doc.data().email, name: doc.data().fname+" "+doc.data().lname, type: "user" }));
+    const usersData = users.map(doc => ({ id: doc.id, email: doc.data().email, name: doc.data().fname + " " + doc.data().lname, type: "user" }));
     const teamsData = teamsBasedOnName.docs.map(doc => ({ id: doc.id, name: doc.data().name, type: "team" }));
 
     response = {
@@ -266,7 +268,7 @@ router.get("/search", async (req, res) => {
 
 
 module.exports = router;
-   
+
